requirements.txt
streamlit>=1.41.0
pandas
streamlit-aggrid
datetime
PyGithub
psycopg2-binary
bcrypt
streamlit-authenticator
google-auth
google-auth-oauthlib
openpyxl
matplotlib
seaborn
plotly
requests>=2.31.0
extra-streamlit-components==0.1.60
google-auth>=2.23.0
google-auth-oauthlib>=1.2.0
authlib>=1.3.2
